final pass on in editor fine tuning

+++ OpenAI/Packages/com.openai.unity/Runtime/FineTuning/CreateFineTuneJobRequest.cs
@@ -12,7 +12,7 @@ namespace OpenAI.FineTuning
             string validationFileId = null,
             string model = null,
             uint epochs = 4,
-            double? batchSize = null,
+            int? batchSize = null,
             double? learningRateMultiplier = null,
             double promptLossWeight = 0.01d,
             bool computeClassificationMetrics = false,
@@ -48,7 +48,7 @@ namespace OpenAI.FineTuning
         public int Epochs { get; set; }
 
         [JsonProperty("batch_size")]
-        public double? BatchSize { get; set; }
+        public int? BatchSize { get; set; }
 
         [JsonProperty("learning_rate_multiplier")]
         public double? LearningRateMultiplier { get; set; }
